backend/requirements.txt
# Multimodal Radiological AI Assistant - Backend Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Deep Learning & Vision
torch>=2.0.0
torchvision>=0.15.0
torchxrayvision==1.2.1

# NLP & Transformers
transformers==4.36.0
sentence-transformers==2.2.2
accelerate==0.25.0
bitsandbytes==0.41.3

# RAG & Search
faiss-cpu==1.7.4
# faiss-gpu==1.7.4  # Uncomment for GPU-accelerated FAISS

# Data Processing
pandas==2.1.4
numpy==1.24.3
scikit-image==0.22.0
Pillow==10.1.0

# Dataset
kagglehub==0.2.5

# Utilities
python-dotenv==1.0.0
